spawning updates
split into multiple classed
