+++ ENGINE/spawn/asset_spawner.hpp
@@ -9,7 +9,7 @@
 #include <nlohmann/json.hpp>
 #include "utils/area.hpp"
 #include "asset/Asset.hpp"
-#include "room/Room.hpp"
+#include "room/room.hpp"
 #include "asset/asset_info.hpp"
 #include "asset/asset_library.hpp"
 #include "asset_spawn_planner.hpp"
