+++ ENGINE/spawn/methods/center_spawner.cpp
@@ -1,0 +1,27 @@
+#include "center_spawner.hpp"
+
+#include "utils/spawn_context.hpp"
+#include "asset_spawn_planner.hpp"
+#include "asset/asset_info.hpp"
+#include "utils/area.hpp"
+#include "spawn_logger.hpp"
+
+void CenterSpawner::spawn(const SpawnInfo& item, const Area* area, SpawnContext& ctx) {
+    if (!item.info || !area) return;
+
+    const int Y_SHIFT = 200;
+    auto center = ctx.get_area_center(*area);
+    center.second -= Y_SHIFT;
+
+    if (ctx.checker().check(item.info, center.first, center.second, ctx.exclusion_zones(), ctx.all_assets(),
+                            item.check_overlap, item.check_min_spacing, false, 5)) {
+        ctx.logger().output_and_log(item.name, item.quantity, 0, 1, 1, "center");
+        return;
+    }
+
+    auto* result = ctx.spawnAsset(item.name, item.info, *area, center.first, center.second, 0, nullptr);
+    int spawned = result ? 1 : 0;
+
+    ctx.logger().progress(item.info, spawned, item.quantity);
+    ctx.logger().output_and_log(item.name, item.quantity, spawned, 1, 1, "center");
+}
