+++ ENGINE/spawn/methods/center_spawner.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+class SpawnInfo;
+class Area;
+class SpawnContext;
+
+class CenterSpawner {
+public:
+    void spawn(const SpawnInfo& item, const Area* area, SpawnContext& ctx);
+};
+
