+++ ENGINE/spawn/methods/distributed_batch_spawner.hpp
@@ -1,0 +1,17 @@
+#pragma once
+
+#include <vector>
+
+class BatchSpawnInfo;
+class Area;
+class SpawnContext;
+
+class DistributedBatchSpawner {
+public:
+    void spawn(const std::vector<BatchSpawnInfo>& items,
+               const Area* area,
+               int spacing,
+               int jitter,
+               SpawnContext& ctx);
+};
+
