+++ MAPS/FORREST/spawn_log.csv
@@ -110,41 +110,41 @@ SKull_pile,0.122,6,49,random,0.579,3,-0.109
 
 
 MAPS/FORREST/rooms
-Davey,1.000,816,816,center,0.065,816,-0.025
-candle,0.009,8,900,random,1.841,3,1.546
-Chest,0.050,2,40,random,0.151,1,0.000
+Davey,1.000,819,819,center,0.065,819,-0.016
+candle,0.010,7,700,random,1.271,3,-0.616
+Chest,0.000,0,170,random,0.021,1,0.000
 lantern,1.000,1,1,exact,0.096,1,0.000
 plant,0.000,0,0,distributed_batch,52.423,141,-38.728
-Skull_1,0.022,432,19540,random,4.398,30,-3.443
-Bone_Pile,0.034,309,9216,random,3.659,16,1.395
-Grave_1,1.000,478,478,distributed_batch,45.075,12,-3.591
-Grave_2,1.000,646,646,distributed_batch,45.132,18,-2.868
+Skull_1,0.022,448,20360,random,4.423,31,0.784
+Bone_Pile,0.034,345,10065,random,3.593,18,-3.186
+Grave_1,0.950,19,20,random,10.562,1,0.000
+Grave_2,1.000,775,775,distributed_batch,45.174,21,-8.837
 Bone_1,0.023,164,7253,random,3.851,12,1.281
-SKull_pile,0.017,239,13990,random,3.892,22,0.407
+SKull_pile,0.017,246,14380,random,3.860,23,-0.747
 Skull,0.962,870,904,random,0.249,284,-0.036
 Bone,0.021,244,11440,random,4.031,16,0.043
 Book,0.965,249,258,random,0.508,18,-0.172
-Fountain,1.000,1,1,center,0.964,1,0.000
-Bench,0.000,0,0,exact,0.000,10,0.000
-flower,0.000,0,0,perimeter,0.005,1,0.000
-Barrel,0.047,7,150,random,595.531,1,0.000
+Fountain,1.000,2,2,center,0.703,2,-0.016
+Bench,0.000,0,0,exact,0.000,1,0.000
+flower,1.000,2,2,center,0.058,2,0.012
+Barrel,1.000,7,7,distributed_batch,98.221,5,-64.671
 Booze_1,0.914,18220,19924,random,0.466,3712,-0.157
 Mushroom_2,0.885,10120,11439,random,1.001,397,-0.208
 Sign,1.000,74,74,center,0.077,74,0.030
 Mushroom_1,0.285,1006,3525,random,0.267,268,1.618
 Booze_2,0.934,12300,13170,random,0.272,3844,0.164
-Key,1.000,11,11,center,0.083,11,-0.022
-car,1.000,2,2,center,0.069,2,-0.005
-flower_1,0.909,60,66,random,3.517,1,0.000
-Grass_FPS_2,1.000,1471419,1471419,distributed_batch,21.675,24237,579.064
-Grass_1_FPS_2,1.000,422,422,distributed_batch,141.587,8,530.987
-bush,0.000,0,0,distributed_batch,48.359,6871,590.250
-Book_Pile_1,0.345,3897,11307,random,1.112,464,-0.072
+Key,1.000,14,14,center,0.079,14,-0.016
+car,1.000,1,1,random,0.056,1,0.000
+flower_1,0.901,64,71,random,1.811,2,-1.316
+Grass_FPS_2,1.000,1472812,1472812,distributed_batch,21.692,24261,25.366
+Grass_1_FPS_2,1.000,476,476,distributed_batch,29.957,8,31.246
+bush,0.000,0,0,distributed_batch,48.444,6895,37.396
+Book_Pile_1,0.339,3903,11497,random,1.111,465,-0.265
 Book_Pile,0.391,3851,9850,random,0.829,472,-0.242
 brick,0.705,31,44,random,1.287,1,0.000
 light,0.000,0,0,distributed_batch,17.115,3887,-1.991
 Tree_Dark,1.000,59,59,exact,0.080,64,-0.080
-Skull_Pile_1,0.009,96,10634,random,3.224,16,0.861
+Skull_Pile_1,0.009,103,11454,random,3.266,17,0.714
 Tree,1.000,107,107,exact,0.071,117,-0.071
 Rock,0.643,709,1103,random,2.363,24,-0.856
 Tree_Darker,1.000,147,147,exact,0.076,161,-0.075
@@ -152,8 +152,8 @@ Big_Rock,1.000,21048,21048,distributed_batch,63.924,9234,-2.473
 Bush_1_FPS_2,1.000,24273,24273,distributed_batch,32.557,17330,-12.594
 Mushroom_3,0.056,65,1161,random,1.666,6,-1.610
 Stump,0.060,31297,525203,random,17.178,2384,-16.588
-Small_Rock,1.000,104569,104569,distributed_batch,42.633,23323,572.739
-Box,0.048,6368,131998,random,4.601,2769,3.782
+Small_Rock,1.000,104629,104629,distributed_batch,42.640,23347,21.520
+Box,0.048,6373,132058,random,4.601,2770,0.622
 noose,0.816,933,1144,random,3.319,25,1.691
 Book_1,0.469,1754,3738,random,0.363,290,0.234
 Bush_FPS_2,1.000,42860,42860,distributed_batch,39.933,17264,-14.205
@@ -162,33 +162,33 @@ Book_2,0.962,127,132,random,0.954,4,-1.165
 Flame_FPS_12,0.002,3,1540,random,1.480,2,0.506
 Big_Rock_Center,1.000,789,789,center,3.445,789,1.871
 Mushroom,0.612,82,134,random,1.109,2,1.125
-flowers,0.000,0,0,distributed_batch,8.926,5651,-0.385
-pretty,0.000,0,0,distributed_batch,49.845,23655,572.289
-table,0.042,4109,98268,random,4.959,1852,2.762
-flower_3,0.850,17,20,random,1.130,1,0.000
-fog,1.000,509182,509182,distributed_batch,57.107,22164,573.411
-flower_5,1.000,1,1,distributed,0.080,1,0.000
-flower_4,1.000,8,8,random,0.583,1,0.000
-flowr_6,0.985,64,65,random,1.952,2,-1.429
+flowers,0.000,0,0,distributed_batch,8.925,5654,5.324
+pretty,0.000,0,0,distributed_batch,49.852,23679,21.596
+table,0.042,4116,98538,random,4.961,1854,2.587
+flower_3,0.000,0,0,perimeter,0.005,1,0.000
+fog,1.000,509323,509323,distributed_batch,57.114,22188,21.112
+flower_5,1.000,1,1,distributed,0.058,1,0.000
+flower_4,1.000,1,1,distributed,0.088,1,0.000
+flowr_6,0.985,64,65,random,3.371,1,0.000
 Lantern,0.007,1,150,random,2.455,1,0.000
 
 
 
 MAPS/FORREST/trails
-light,0.000,0,0,distributed_batch,20.692,36776,-16.853
+light,0.000,0,0,distributed_batch,20.691,36806,-10.566
 flower,1.000,91164,91164,distributed_batch,39.899,13432,-25.662
-candle,1.000,1,1,distributed_batch,10.180,1,0.000
-flower_1,1.000,80392,80392,distributed_batch,14.408,17783,-0.039
-Grass_FPS_2,1.000,1345919,1345919,distributed_batch,18.264,34641,-2.385
-Grass_1_FPS_2,1.000,1741302,1741302,distributed_batch,25.322,34640,-1.905
-bush,0.000,0,0,distributed_batch,47.188,8036,5.031
+candle,1.000,2,2,random,0.073,2,-0.012
+flower_1,1.000,80443,80443,distributed_batch,14.408,17796,-1.839
+Grass_FPS_2,1.000,1347205,1347205,distributed_batch,18.264,34671,0.723
+Grass_1_FPS_2,1.000,1742588,1742588,distributed_batch,25.322,34670,0.600
+bush,0.000,0,0,distributed_batch,47.229,8066,9.686
 Bush_1_FPS_2,1.000,29604,29604,distributed_batch,32.248,26604,-6.516
 Bush_FPS_2,1.000,46600,46600,distributed_batch,39.099,26604,-8.368
-Small_Rock,1.000,105643,105643,distributed_batch,43.214,33401,-12.142
+Small_Rock,1.000,105715,105715,distributed_batch,43.214,33431,-8.576
 car,1.000,240,240,distributed_batch,54.595,1173,-20.413
-pretty,0.000,0,0,distributed_batch,48.632,28103,-10.236
-fog,1.000,448537,448537,distributed_batch,54.967,26672,-9.503
+pretty,0.000,0,0,distributed_batch,48.632,28133,-6.399
+fog,1.000,448690,448690,distributed_batch,54.966,26702,-5.014
 Big_Rock,1.000,24113,24113,distributed_batch,67.330,12717,12.582
 Axe,1.000,199,199,distributed_batch,27.841,23,0.293
 lantern,0.066,634,9622,random,0.789,405,-0.335
-flowers,0.000,0,0,distributed_batch,5.878,5046,-0.574
+flowers,0.000,0,0,distributed_batch,5.879,5063,0.242
